converted main display to div from table
Made the main scoreboard display easily to manipulate by converting the table that was there previous to div and using a flex grid for placement of the respective divs. I've swapped around the player name and current score placement I think the time and should be the two most prominent value. The total display is now dynamically the same height as the viewport, can be used to snap to position when play button pressed. Next step is to size and style.